Move genswagger to public API again
Fixes #1162
