+++ protoc-gen-swagger/BUILD.bazel
@@ -9,7 +9,7 @@ go_library(
     deps = [
         "//internal/codegenerator:go_default_library",
         "//protoc-gen-grpc-gateway/descriptor:go_default_library",
-        "//protoc-gen-swagger/internal/genswagger:go_default_library",
+        "//protoc-gen-swagger/genswagger:go_default_library",
         "@com_github_golang_glog//:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
         "@io_bazel_rules_go//proto/wkt:compiler_plugin_go_proto",
