+++ protoc-gen-swagger/genswagger/BUILD.bazel
@@ -12,7 +12,7 @@ go_library(
         "template.go",
         "types.go",
     ],
-    importpath = "github.com/grpc-ecosystem/grpc-gateway/protoc-gen-swagger/internal/genswagger",
+    importpath = "github.com/grpc-ecosystem/grpc-gateway/protoc-gen-swagger/genswagger",
     deps = [
         "//internal:go_default_library",
         "//protoc-gen-grpc-gateway/descriptor:go_default_library",
