+++ protoc-gen-swagger/genswagger/helpers_go111_old.go
